// src/store/account/types.ts
export const LOGIN_REQUEST = 'LOGIN_REQUEST';
export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
export const LOGIN_FAILURE = 'LOGIN_FAILURE';

export const LOG_OUT = 'LOG_OUT';

//action includes Type and Payload
export interface AuthenicatedUser {
    _id: string;
    first_name: string;
    last_name: string;
    email: string;
    avatar: string;
}
interface LoginRequest {
    type: typeof LOGIN_REQUEST;
    payload: {
        email: string;
        password: string;
    }
}
interface LoginSuccess {
    type: typeof LOGIN_SUCCESS;
    payload: {
        token: string;
    }
}
interface LoginFailure {
    type: typeof LOGIN_FAILURE;
    payload: {
        error: string
    }
}
interface LogOut {
    type: typeof LOG_OUT
}

//store
export interface AccountState {
    user: AuthenicatedUser | null;
    loading: boolean;
    error: string | null;
    token: string | null;
}

//action type
export type AccountActionTypes =
    | LoginRequest
    | LoginSuccess
    | LoginFailure
    | LogOut
    | LogOut;